requirements.txt
numpy~=1.22.3
Pillow~=9.1.0
keras~=2.8.0
opencv-python~=4.5.5.64
pydantic~=1.9.0
fastapi~=0.75.1
uvicorn~=0.17.6